chore: prepare-strapi to prepare-navigation, update readme
