+++ server/strapi.ts
@@ -1,5 +1,7 @@
+import { writeFileSync } from "fs";
+
 const fetchUrl = "https://goteleport.com/api/data/navigation";
-export const fetchData = async () => {
+const fetchData = async () => {
   const abortController = new AbortController();
   try {
     // Timeout all strapi requests after 10 seconds
@@ -31,5 +33,37 @@ export const fetchData = async () => {
       console.error("Unexpected error:", err);
     }
     return [];
+  }
+};
+
+export const generateData = async ({
+  navPath,
+  eventPath,
+}: {
+  navPath: string;
+  eventPath: string;
+}) => {
+  let data = undefined;
+  try {
+    data = await fetchData();
+    if (!data) return;
+  } catch (error) {
+    console.error("No data returned");
+    return;
+  }
+
+  if (data) {
+    try {
+      writeFileSync(navPath, JSON.stringify(data.navbardata));
+      console.log("Writing header data to file: ", navPath);
+    } catch (error) {
+      console.error("Error writing footer data to file:", error);
+    }
+    try {
+      writeFileSync(eventPath, JSON.stringify(data.eventsdata));
+      console.log("Writing event data to file: ", eventPath);
+    } catch (error) {
+      console.error("Error writing footer data to file:", error);
+    }
   }
 };
